Refactor: Enforce semantic module boundaries
Restricted visibility of `Assembler`, `AssembledStatement` and related types to `pub(crate)`
to enforce the public API boundary of the semantic analysis module.

- Moved internal integration tests to `src/semantic/assembler_tests.rs`
- Updated `src/semantic/mod.rs` to stop exporting internal types
- Restricted visibility of constants and helper functions
- Fixed doctests to respect new visibility
- Cleaned up unused imports and dead code warnings
