+++ src/semantic/mod.rs
@@ -32,6 +32,8 @@
 //! without changing the semantic meaning.
 
 pub mod assembler;
+#[cfg(test)]
+mod assembler_tests;
 pub(crate) mod assembly_model;
 pub(crate) mod conversion;
 pub mod expressions;
@@ -42,10 +44,9 @@ pub(crate) mod statements;
 mod types;
 
 pub use crate::morphology::{DisambiguationContext, analyze_article, disambiguate, resolve_best};
-pub use assembler::{Assembler, AssemblyError};
-pub use assembly_model::{
-    AssembledStatement, Constituent, Literal, ParticipleConstituent, VerbConstituent,
-};
+pub(crate) use assembler::Assembler;
+pub use assembler::AssemblyError;
+pub(crate) use assembly_model::{AssembledStatement, Constituent, Literal};
 pub use model::*;
 pub use resolver::*;
 pub use types::*;
@@ -322,7 +323,7 @@ fn extract_block_statements(stmt: &Statement) -> Option<&Vec<Statement>> {
 }
 
 /// Analyze a single statement using the slot-based assembler
-pub fn assemble_statement(stmt: &Statement) -> Result<AssembledStatement, GlossaError> {
+pub(crate) fn assemble_statement(stmt: &Statement) -> Result<AssembledStatement, GlossaError> {
     let mut asm = Assembler::new();
     asm.set_query(stmt.is_query());
     asm.set_propagate(stmt.is_propagate());
